estudo-flask/AutenticacaoUserRESTcomJWT/requirements.txt
aniso8601==9.0.1
autopep8==1.5.7
click==8.0.1
Flask==2.0.1
Flask-JWT-Extended==4.2.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycodestyle==2.7.0
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.16
toml==0.10.2
Werkzeug==2.0.1
